Moved typescript tests and tsconfig to own folder
