+++ package.json
@@ -16,9 +16,9 @@
     "lint": "eslint --ext .js lib test",
     "test": "npm run test:2.6 && npm run test:2.7 && npm run test:3",
     "tests": "node test/shell.js",
-    "pretest:2.6": "npm install --save-dev @vue/test-utils@1.1.1 vue@2.6 vue-template-compiler@2.6 @vue/component-compiler-utils",
+    "pretest:2.6": "npm install --save-dev @vue/test-utils@1.1.1 vue@2.6 vue-template-compiler@2.6 @vue/component-compiler-utils typescript@5",
     "test:2.6": "npm run tests",
-    "posttest:2.6": "npm uninstall @vue/test-utils vue vue-template-compiler @vue/component-compiler-utils",
+    "posttest:2.6": "npm uninstall @vue/test-utils vue vue-template-compiler @vue/component-compiler-utils typescript",
     "pretest:2.7": "npm install --save-dev @vue/test-utils@1.1.1 vue@2 typescript@5",
     "test:2.7": "npm run tests",
     "posttest:2.7": "npm uninstall @vue/test-utils vue typescript",
