+++ test/options/options-test.js
@@ -10,11 +10,6 @@ describe('vue-esm-loader with custom options', function() {
 			let filePath = `./files/${id}`;
 			return (await import(filePath)).default;
 		};
-		this.semver = function(expr) {
-			if (!semver.satisfies(version, expr)) {
-				this.skip();
-			}
-		};
 	});
 
 	it('imports .vue files', async function() {
@@ -24,27 +19,6 @@ describe('vue-esm-loader with custom options', function() {
 		expect(Component.props).to.eql(['vue']);
 
 	});
-
-	it('imports .vue files that use typescript and composition api', async function() {
-		this.semver('>=2.7');
-
-		const Component = await this.require('ts-component-composition.vue');
-		expect(Component.render).to.be.ok;
-		expect(Component.props).to.have.all.keys('message');
-		expect(Component.props.message.required).to.be.true;
-
-	});
-
-	it('imports .vue files that use typescript and options api', async function() {
-		this.semver('>=2.7');
-
-		const Component = await this.require('ts-component-options.vue');
-		expect(Component.render).to.be.ok;
-		expect(Component.props).to.have.all.keys('message');
-		expect(Component.props.message.required).to.be.true;
-
-	});
-
 	it('imports .vuex files', async function() {
 
 		const Component = await this.require('component.vuex');
