+++ test/typescript/.mocharc.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+	'experimental-loader': './typescript-loader.js',
+	'no-warnings': true,
+};
