+++ test/typescript/files/ts-component-composition.vue
