+++ test/typescript/files/ts-component-options.vue
